pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-view-3d-reconstruction"
version = "1.0.0"
description = "Modern multi-view 3D reconstruction using PyTorch and OpenCV"
authors = [
    {name = "Multi-view 3D Reconstruction Team", email = "team@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "open3d>=0.17.0",
    "pytorch3d>=0.7.4",
    "trimesh>=3.23.0",
    "kornia>=0.7.0",
    "opencv-python>=4.8.0",
    "nerfstudio>=0.3.0",
    "pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "imageio>=2.31.0",
    "streamlit>=1.25.0",
    "gradio>=3.40.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "pytest>=7.4.0",
    "pre-commit>=3.3.0",
]
tracking = [
    "wandb>=0.15.0",
    "mlflow>=2.5.0",
]
serving = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.scripts]
train-sfm = "scripts.train:main"
evaluate-sfm = "scripts.evaluate:main"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
